requirements.txt
adjustText==0.7.3
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.1.0
drf-writable-nested==0.6.3
drf-yasg==1.20.0
Faker==12.0.1
fonttools==4.33.3
groups==0.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
many==0.6.9
MarkupSafe==2.1.1
matplotlib==3.5.1
mixer==7.2.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
patsy==0.5.2
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
self==2020.12.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
statsmodels==0.13.2
threadpoolctl==3.1.0
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
